+++ Clean-Restic.ps1
@@ -24,12 +24,12 @@
         Will simulate removing of V Rising snapshots
     .NOTES
         Name           : Clean-Restic
-        Version        : 2.0
+        Version        : 2.1
         Created by     : Chucky2401
         Date Created   : 30/06/2022
         Modify by      : Chucky2401
-        Date modified  : 03/01/2023
-        Change         : Settings / Environment / Localized
+        Date modified  : 11/05/2023
+        Change         : Visual enhancement
     .LINK
         https://github.com/Chucky2401/Restic-Scripts/blob/main/README.md#clean-restic
 #>
@@ -60,43 +60,51 @@ BEGIN {
     #---------------------------------------------------------[Initialisations]--------------------------------------------------------
 
     #Set Error Action to Silently Continue
-    $ErrorActionPreference = "Stop"
+    $global:ErrorActionPreference = "Stop"
+    $global:DebugPreference       = 'SilentlyContinue'
+    If ($PSBoundParameters['Debug']) {
+        $global:DebugPreference = 'Continue'
+    }
 
     Update-FormatData -AppendPath "$($PSScriptRoot)\inc\format\ResticControl.format.ps1xml"
+    $PSStyle.Progress.MaxWidth = ($Host.UI.RawUI.WindowSize.Width)-5
+
     Import-LocalizedData -BindingVariable "Message" -BaseDirectory "local" -FileName "Clean-Restic.psd1"
 
-    Import-Module -Name ".\inc\modules\Tjvs.Message"
-    Import-Module -Name ".\inc\modules\Tjvs.Process"
-    Import-Module -Name ".\inc\modules\Tjvs.Restic"
     Import-Module -Name ".\inc\modules\Tjvs.Settings"
+    Import-Module -Name ".\inc\modules\Tjvs.Message", ".\inc\modules\Tjvs.Process", ".\inc\modules\Tjvs.Restic"
+
+    #Set-PowerShellUICulture en-US
 
     #-----------------------------------------------------------[Functions]------------------------------------------------------------
 
     #----------------------------------------------------------[Declarations]----------------------------------------------------------
 
     # Settings
-    If (-not (Test-Path ".\conf\settings.json")) {
-        Write-Warning $Message.NoSetFile
-        Write-Host $Message.PleaseAnswer
-        
-        New-Settings -RootPath $PSScriptRoot
-    }
+    #If (-not (Test-Path ".\conf\settings.json")) {
+    #    Write-Warning $Message.NoSetFile
+    #    Write-Host $Message.PleaseAnswer
+    #    
+    #    New-Settings -RootPath $PSScriptRoot
+    #}
     
-    $oSettings = Get-Settings -File ".\conf\settings.json"
+    #$oSettings = Get-Settings -File ".\conf\settings.json"
 
     ## Default settings
-    If ($SnapshotToKeep -eq 0) {
-        $SnapshotToKeep = $oSettings.Snapshots.ToKeep
+    If ($PSBoundParameters.ContainsKey('SnapshotToKeep') -eq $False) {
+        $SnapshotToKeep = $global:settings.Snapshots.ToKeep
     }
 
     # Restic Info
     ## Envrinoment variable
-    Set-Environment -Settings $oSettings
+    #Set-Environment -Settings $oSettings
 
     ## Common restic to use
-    $sFilter = "--tag `"$sGame`""
+    $sFilter = "--tag `"$Game`""
+    $messageTagFilter = ""
     If ($TagFilter -ne "") {
-        $sFilter += ",`"$($TagFilter)`""
+        $sFilter = "--tag `"$Game,$($TagFilter)`""
+        $messageTagFilter = $Message.Oth_MessageFilter -f $TagFilter
     }
     
     # Logs
@@ -109,6 +117,7 @@ BEGIN {
 
     # Init Var
     $oDataBefore = $null
+    $cntDetails  = 1
     
     #-----------------------------------------------------------[Execution]------------------------------------------------------------
 
@@ -122,14 +131,18 @@ BEGIN {
     Write-CenterText "*          Start $(Get-Date -Format 'HH:mm')          *" $sLogFile
     Write-CenterText "*                               *" $sLogFile
     Write-CenterText "*********************************" $sLogFile
-    ShowLogMessage "OTHER" "" -sLogFile ([ref]$sLogFile)
+    ShowLogMessage -type "OTHER" -message "" -sLogFile ([ref]$sLogFile)
 
-    If (-not $oSettings.Global.Stats) {
+    If (-not $global:settings.Global.Stats) {
         Write-Warning $Message.Warn_StatsDisable
-        ShowLogMessage "OTHER" "" -sLogFile ([ref]$sLogFile)
+        ShowLogMessage -type "OTHER" -message "" -sLogFile ([ref]$sLogFile)
         $NoStats = $True
     }
 
+    ## Demo purpose only!
+    #$NoStats = $True
+    #$NoDelete = $True
+    ## Demo purpose only!
 }
 
 PROCESS {
@@ -156,20 +169,29 @@ PROCESS {
             PAUSE
             exit 1
         }
+
+        $numberSnapshotsTotal = $jsResultRestic.Count
+        $numberSnapshotsToRemove = ($jsResultRestic | Sort-Object time | Select-Object -SkipLast $SnapshotToKeep).Count
         
         ShowLogMessage -type "OTHER" -message "" -sLogFile ([ref]$sLogFile)
         
-        ShowLogMessage -type "INFO" -message $Message.Inf_DelSnaps -variable $($sGame),$($SnapshotToKeep) -sLogFile ([ref]$sLogFile)
+        If ($numberSnapshotsTotal -eq $numberSnapshotsToRemove) {
+            ShowLogMessage -type "INFO" -message $Message.Inf_DelSnapsAll -variable $($sGame),$messageTagFilter -sLogFile ([ref]$sLogFile)
+        } Else {
+            ShowLogMessage -type "INFO" -message $Message.Inf_DelSnaps -variable $numberSnapshotsToRemove,$numberSnapshotsTotal,$($sGame),$messageTagFilter -sLogFile ([ref]$sLogFile)
+        }
         $jsResultRestic | Sort-Object time | Select-Object -SkipLast $SnapshotToKeep | ForEach-Object {
+            $iPercentComplete = [Math]::Round(($cntDetails/$numberSnapshotsToRemove)*100,2)
             $sSnapshotId = $PSItem.short_id
+
+            Write-Progress -Activity $($Message.Prg_Activity -f $($sGame), $($cntDetails), $($numberSnapshotsToRemove), $($iPercentComplete)) -PercentComplete $iPercentComplete -Status $($Message.Prg_Status -f $($sSnapshotId))
+
             If (!$NoDelete) {
                 $oResticProcess = Start-Command -Title "Restic - Forget $($sSnapshotId)" -FilePath restic -ArgumentList "forget --tag `"$sGame`" $sSnapshotId"
             
                 If ($oResticProcess.ExitCode -eq 0) {
                     $aResultDelete     = $oResticProcess.stdout.Split("`n") | Where-Object { $PSItem -ne "" }
                     $aSnapshotRemoved += [PSCustomObject]@{ SnapshotId = $sSnapshotId ; Detail = [String]::Join("//", $aResultDelete) }
-    
-                    ShowLogMessage -type "SUCCESS" -message $Message.Suc_DelSnaps -variable $($sSnapshotId) -sLogFile ([ref]$sLogFile)
                 } Else {
                     $aResultDelete          = $oResticProcess.stderr.Split("`n") | Where-Object { $PSItem -ne "" }
                     $aSnapshotStillPresent += [PSCustomObject]@{ SnapshotId = $sSnapshotId ; Detail = [String]::Join("//", $aResultDelete) }
@@ -182,10 +204,13 @@ PROCESS {
                     }
                 }
             } Else {
-                ShowLogMessage -type "DEBUG" -message $Message.Dbg_DelSnaps -variable $($sSnapshotId) -sLogFile ([ref]$sLogFile)
+                ShowLogMessage -type "OTHER" -message $Message.Dbg_DelSnaps -variable $($sSnapshotId) -sLogFile ([ref]$sLogFile)
+                #Start-Sleep -Seconds 2
                 $aSnapshotRemoved += $sSnapshotId
             }
+            $cntDetails++
         }
+        Write-Progress -Activity $Message.Prg_Complete -Completed
     
         If (!$NoDelete) {
             If ($aSnapshotStillPresent.Count -ge 1) {
@@ -194,7 +219,8 @@ PROCESS {
                 ShowLogMessage -type "SUCCESS" -message $Message.Suc_SumDel -variable $($aSnapshotRemoved.Count) -sLogFile ([ref]$sLogFile)
             }
         } Else {
-            ShowLogMessage -type "DEBUG" -message $Message.Dbg_SumDel -variable $($aSnapshotRemoved.Count) -sLogFile ([ref]$sLogFile)
+            ShowLogMessage -type "OTHER" -message $Message.Dbg_SumDel -variable $($aSnapshotRemoved.Count) -sLogFile ([ref]$sLogFile)
+            #ShowLogMessage -type "SUCCESS" -message $Message.Suc_SumDel -variable $($aSnapshotRemoved.Count) -sLogFile ([ref]$sLogFile)
         }
     
         ShowLogMessage -type "OTHER" -message "" -sLogFile ([ref]$sLogFile)